upload/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
boto3==1.34.64
botocore==1.34.64
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
distro==1.9.0
eventlet==0.36.1
Flask==3.0.2
Flask-SocketIO==5.3.6
Flask-WTF==1.2.1
fonttools==4.50.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
numpy==1.26.4
openai==1.14.1
packaging==24.0
pandas==2.2.1
pillow==10.2.0
pydantic==2.6.4
pydantic_core==2.16.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.1
pytz==2024.1
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.16
Werkzeug==3.0.1
WTForms==3.1.2
paho-mqtt==1.6.1

